+++ leaf-core/src/main/java/com/sankuai/inf/leaf/exception/InitException.java
@@ -1,4 +1,4 @@
-package com.sankuai.inf.leaf.server.exception;
+package com.sankuai.inf.leaf.exception;
 
 public class InitException extends Exception{
     public InitException(String msg) {
