+++ leaf-server/src/main/resources/leaf-application.xml
@@ -4,6 +4,6 @@
        xmlns:context="http://www.springframework.org/schema/context"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <bean id="SnowflakeService" class="com.sankuai.inf.leaf.server.service.SnowflakeService"/>
-    <bean id="SegmentService" class="com.sankuai.inf.leaf.server.service.SegmentService"/>
+    <bean id="SnowflakeService" class="com.sankuai.inf.leaf.service.SnowflakeService"/>
+    <bean id="SegmentService" class="com.sankuai.inf.leaf.service.SegmentService"/>
 </beans>
